Added lsp_signature.nvim

+++ lua/core/pack.lua
@@ -26,58 +26,42 @@ function Packer:load_packer()
         api.nvim_command("packadd packer.nvim")
         packer = require("packer")
     end
-    packer.init(
-        {
-            compile_path = packer_compiled,
-            git = {clone_timeout = 120},
-            disable_commands = true
-        }
-    )
+    packer.init({
+        compile_path = packer_compiled,
+        git = {clone_timeout = 120},
+        disable_commands = true
+    })
     packer.reset()
     local use = packer.use
     self:load_plugins()
     use {"wbthomason/packer.nvim", opt = true}
-    for _, repo in ipairs(self.repos) do
-        use(repo)
-    end
+    for _, repo in ipairs(self.repos) do use(repo) end
 end
 
 function Packer:init_ensure_plugins()
-    local packer_dir =
-        data_path ..
-        "pack" .. global.path_sep .. "packer" .. global.path_sep .. "opt" .. global.path_sep .. "packer.nvim"
+    local packer_dir = data_path .. "pack" .. global.path_sep .. "packer" ..
+                           global.path_sep .. "opt" .. global.path_sep ..
+                           "packer.nvim"
     local state = uv.fs_stat(packer_dir)
     if not state then
-        local cmd = "!git clone https://github.com/wbthomason/packer.nvim " .. packer_dir
+        local cmd = "!git clone https://github.com/wbthomason/packer.nvim " ..
+                        packer_dir
         api.nvim_command(cmd)
-        uv.fs_mkdir(
-            data_path .. "lua",
-            511,
-            function()
-                assert("make compile path dir failed")
-            end
-        )
+        uv.fs_mkdir(data_path .. "lua", 511,
+                    function() assert("make compile path dir failed") end)
         self:load_packer()
         packer.install()
     end
 end
 
-local plugins =
-    setmetatable(
-    {},
-    {
-        __index = function(_, key)
-            if not packer then
-                Packer:load_packer()
-            end
-            return packer[key]
-        end
-    }
-)
+local plugins = setmetatable({}, {
+    __index = function(_, key)
+        if not packer then Packer:load_packer() end
+        return packer[key]
+    end
+})
 
-function plugins.ensure_plugins()
-    Packer:init_ensure_plugins()
-end
+function plugins.ensure_plugins() Packer:init_ensure_plugins() end
 
 function plugins.convert_compile_file()
     local lines = {}
@@ -87,22 +71,16 @@ function plugins.convert_compile_file()
         lnum = lnum + 1
         if lnum > 15 then
             lines[#lines + 1] = line .. "\n"
-            if line == "END" then
-                break
-            end
+            if line == "END" then break end
         end
     end
     table.remove(lines, #lines)
     if fn.isdirectory(data_path .. "lua") ~= 1 then
         os.execute("mkdir -p " .. data_path .. "lua")
     end
-    if fn.filereadable(compile_to_lua) == 1 then
-        os.remove(compile_to_lua)
-    end
+    if fn.filereadable(compile_to_lua) == 1 then os.remove(compile_to_lua) end
     local file = io.open(compile_to_lua, "w")
-    for _, line in ipairs(lines) do
-        file:write(line)
-    end
+    for _, line in ipairs(lines) do file:write(line) end
     file:close()
 
     os.remove(packer_compiled)
@@ -117,7 +95,8 @@ function plugins.load_compile()
     if fn.filereadable(compile_to_lua) == 1 then
         require("_compiled")
     else
-        assert("Missing packer compile file Run PackerCompile Or PackerInstall to fix")
+        assert(
+            "Missing packer compile file Run PackerCompile Or PackerInstall to fix")
     end
     vim.cmd [[command! PackerCompile lua require('core.pack').auto_compile()]]
     vim.cmd [[command! PackerInstall lua require('core.pack').install()]]
